Ground Work for PKK
This patch is the first of several commits that enable the use of
Intel's PKU feature for kernel memory. We refer to this new feature as:
"Protection Keys for Kernel (PKK)".

All pages are marked as user pages (U/S bit=1), and pages that used to be user
are modified to have PKEY = 8 by default (instead of PKEY = 0).
In IskiOS, we plan to have PKEYs 8-15 for user-space
and 0-7 for kernel pages. SMEP and SMAP should be disabled when PKK is enabled
(nosmep,nosmap in boot parameters). The SMEP feature will be implemented
as an LLVM pass which instruments return instructions and indirect branches.
The SMAP feature will be enforced when we disallow kernel to access pages with
PKEY 8-15 (not yet implemented).

This patch was created by Mohammad Hedayati and Spyridoula Gravani.