PKRU Save/Restore Optimization
This commit optimizes the PKRU save/restore on kernel entry/exit respectively.
Specifically, since the wrpkru instruction is quite expensive, we avoid overwriting the contents of
the pkru register when the value to be written is the same as the current value of pkru register.
This is common when processes don't use the PKU feature, or if they don't modify the value for
the lower 8 keys which belong to the kernel.

This patch was created by Mohammad Hedayati and Spyridoula Gravani.