PKK Safe WRPKRUs and Config Option
This commit adds a config option which enables Protection Keys for Kernel (PKK)
(option CONFIG_PKK). PKK is by default disabled.
The patch also adds instrumentation on wrpkru instructions
in the kernel to ensure that they are never executed from code running
in user mode. This ensures that any PKK isolation mechanism cannot be bypassed
by simply setting the contents of %rax and jumping to a wrpkru instruction
from user code.

Note: temporarily, we use a nop instruction instead of a trap when user mode code
executes a wrpkru. This needs to be fixed in future commit.