PKRU Save/Restore Fix
This patch fixes a bug in saving/restoring the pkru register.
The bug is the following: if the processor attempts to perform a rdpkru/wrpkru instruction
before the PKU feature has been enabled, the kernel will hang.
This happens in particular at early boot stage since the PKU feature of secondary CPUs
has not been enabled yet.

The patch was created by Mohammad Hedayati and Spyridoula Gravani.